feat(policy): ADR-063 Phase 9 — consolidate selector matching (D10)
suppression.py's Suppression and reclassify.py's ReclassifyRule now share
one selector-matching primitive, abicheck/policy/selectors.py's new
SelectorSet, instead of two independently-maintained copies of the same
fnmatch/regex/namespace-glob grammar (symbol/symbol_pattern/type_pattern/
member_name/namespace/entity_namespace/cause_namespace/source_location/
change_kind/binding/finding_id/expires).

The leaf is genuinely dependency-free — zero import of checker_types.py/
suppression.py/reclassify.py/policy_file.py/finding_identity.py, enforced
by a new scripts/check_architecture.py `selector-leaf-purity` gate scoped
to this one file — which is what lets reclassify.py import it statically.
Before this phase, reclassify.py built a Suppression instance purely for
its selector grammar, resolved through a runtime importlib.import_module
call to dodge the import cycle a static Suppression import would have
closed (policy_file -> reclassify -> suppression -> checker_types ->
policy_file). That workaround, and its module-docstring justification, are
both gone: neither module needs to import the other any more.

The finding_id matcher (Suppression-only) is a plain string-equality
comparison against an already-computed report_canonical_finding_id(change)
the caller passes in, not a finding_identity.py import from inside the
leaf — the same upward-dependency fix Phase 2 already established for
model/identity.py.

abicheck/policy/selectors_namespace_glob.py holds the namespace-glob
compilation machinery, split out purely to keep both files under the
architecture gate's 800-line ceiling (mechanical extraction).

Selector-matching behavior is unchanged: every existing suppression.py/
reclassify.py selector test passes unmodified against the shared matcher.
New tests/test_selectors.py and tests/test_selectors_namespace_glob.py
exercise the leaf modules directly; tests/test_architecture_check.py gains
cases for the new leaf-purity gate. pyproject.toml's mutation-testing
`only_mutate` list, .github/workflows/mutation.yml's path filters, and
tests/test_mutation_workflow_contract.py's `_ACCEPTED_KILL_LOSS` ledger
all follow the moved matching logic so mutation coverage doesn't silently
regress.

Updates ADR-063's own Status paragraph, the implementation plan's Phase 9
section, and docs/contribute/adr/index.md to record Phase 9 complete.
